+++ resources/js/components/Checkout/Checkout.vue
@@ -26,13 +26,20 @@ export default {
 
         this.steps = this.config.checkout_steps[window.config.store_code] ?? this.config.checkout_steps['default']
         this.setupHistory()
-        this.setCheckoutCredentialsFromDefaultUserAddresses()
-        this.getShippingMethods()
-        this.getTotalsInformation()
-        this.$root.$emit('checkout-step', 1)
+        this.setupCheckout()
     },
 
     methods: {
+        async setupCheckout() {
+            await this.refreshUser(false)
+
+            this.setCheckoutCredentialsFromDefaultUserAddresses()
+            this.getShippingMethods()
+            this.getTotalsInformation()
+
+            this.$root.$emit('checkout-step', 1)
+        },
+
         async getShippingMethods() {
             let responseData = await this.magentoCart('post', 'estimate-shipping-methods', {
                 address: {
@@ -193,6 +200,11 @@ export default {
                     method: this.checkout.payment_method,
                 },
             })
+
+            this.$root.$emit('checkout-payment-selected', {
+                method: this.checkout.payment_method,
+            })
+
             this.getTotalsInformation()
         },
 
@@ -203,22 +215,31 @@ export default {
             }
 
             try {
-                let response = await this.magentoCart('post', 'payment-information', {
-                    billingAddress: this.billingAddress,
-                    shippingAddress: this.shippingAddress,
-                    email: this.user?.email ? this.user.email : this.$root.guestEmail,
-                    paymentMethod: {
-                        method: this.checkout.payment_method,
-                        extension_attributes: {
-                            agreement_ids: this.checkout.agreement_ids,
-                        },
+                this.$root.$emit('before-checkout-payment-saved', {
+                    order: {
+                        payment_method_code: this.checkout.payment_method,
                     },
                 })
+
+                let response = {}
+                if (!window.app.checkout?.preventOrder) {
+                    response = await this.magentoCart('post', 'payment-information', {
+                        billingAddress: this.billingAddress,
+                        shippingAddress: this.shippingAddress,
+                        email: this.user?.email ? this.user.email : this.$root.guestEmail,
+                        paymentMethod: {
+                            method: this.checkout.payment_method,
+                            extension_attributes: {
+                                agreement_ids: this.checkout.agreement_ids,
+                            },
+                        },
+                    })
+                }
                 // response.data = orderId
 
                 this.$root.$emit('checkout-payment-saved', {
                     order: {
-                        id: response.data,
+                        id: response?.data,
                         payment_method_code: this.checkout.payment_method,
                     },
                 })