+++ src/Commands/IndexProductsCommand.php
@@ -51,7 +51,7 @@ class IndexProductsCommand extends ElasticsearchIndexCommand
                     ->pluck('position', 'category_id');
 
                 $productQuery = $productModel::selectOnlyIndexable()
-                    ->with('categoryProducts')
+                    ->with(['categoryProducts', 'reviewSummary'])
                     ->withEventyGlobalScopes('index.product.scopes')
                     ->withExists('options AS has_options');