+++ src/Models/Product.php
@@ -6,6 +6,8 @@ use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Casts\Attribute;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\Relations\HasMany;
+use Illuminate\Database\Eloquent\Relations\HasOne;
+use Illuminate\Database\Eloquent\Relations\HasOneThrough;
 use Rapidez\Core\Casts\Children;
 use Rapidez\Core\Casts\CommaSeparatedToArray;
 use Rapidez\Core\Casts\CommaSeparatedToIntegerArray;
@@ -117,12 +119,30 @@ class Product extends Model
             );
     }
 
+    public function reviewSummary(): HasOne
+    {
+        return $this->hasOne(
+            config('rapidez.models.product_review_summary', Rapidez\Core\Models\ProductReviewSummary::class),
+            'entity_pk_value'
+        );
+    }
+
     public function rewrites(): HasMany
     {
         return $this
             ->hasMany(config('rapidez.models.rewrite'), 'entity_id')
             ->withoutGlobalScope('store')
             ->where('entity_type', 'product');
+    }
+
+    public function parent(): HasOneThrough
+    {
+        return $this->hasOneThrough(
+            config('rapidez.models.product'),
+            config('rapidez.models.product_link'),
+            'product_id', 'entity_id',
+            'entity_id', 'parent_id'
+        )->withoutGlobalScopes();
     }
 
     public function scopeByIds(Builder $query, array $productIds): Builder