+++ src/Models/Quote.php
@@ -12,6 +12,13 @@ class Quote extends Model
 
     protected $primaryKey = 'entity_id';
 
+    protected $guarded = [];
+
+    protected $casts = [
+        'items'       => QuoteItems::class,
+        'cross_sells' => CommaSeparatedToIntegerArray::class,
+    ];
+
     protected static function booting()
     {
         static::addGlobalScope(new IsActiveScope);