+++ src/Rapidez.php
@@ -5,7 +5,9 @@ namespace Rapidez\Core;
 use Illuminate\Routing\RouteAction;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Collection;
+use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Event;
 use Rapidez\Core\Models\Store;
 
 class Rapidez
@@ -109,6 +111,11 @@ class Rapidez
         config()->set('rapidez.website_code', $store['website_code']);
         config()->set('rapidez.group', $store['group_id']);
         config()->set('rapidez.root_category_id', $store['root_category_id']);
+        config()->set('frontend.base_url', url('/'));
+
+        App::setLocale(strtok(Rapidez::config('general/locale/code', 'en_US'), '_'));
+
+        Event::dispatch('rapidez:store-set', [$store]);
     }
 
     public function withStore(Store|array|callable|int|string $store, callable $callback, ...$args)