+++ CHANGELOG.md
@@ -1,6 +1,67 @@
 # Changelog 
 
-[Unreleased changes](https://github.com/rapidez/core/compare/1.6.1...master)
+[Unreleased changes](https://github.com/rapidez/core/compare/1.10.1...master)
+## [1.10.1](https://github.com/rapidez/core/releases/tag/1.10.1) - 2024-01-30
+
+### Fixed
+
+ - Fix slot scope "limit" to "size" (https://github.com/rapidez/core/pull/412)
+
+## [1.10.0](https://github.com/rapidez/core/releases/tag/1.10.0) - 2024-01-29
+
+### Added
+
+ - Add new storecode if directive (https://github.com/rapidez/core/pull/411)
+ - Add product parent (https://github.com/rapidez/core/pull/404)
+ - Improve flexibility of additional search queries in autocomplete (https://github.com/rapidez/core/pull/407)
+
+### Changed
+
+- Always set step in checkout success to success step (https://github.com/rapidez/core/pull/403)
+- Refresh user before retrieving addresses (https://github.com/rapidez/core/pull/345)
+
+### Fixed
+
+ - Fix usage of whitespace tokenizer (https://github.com/rapidez/core/pull/410)
+
+
+## [1.9.0](https://github.com/rapidez/core/releases/tag/1.9.0) - 2024-01-24
+
+### Changed
+
+- Use the configured locale from Magento (#409)
+
+## [1.8.0](https://github.com/rapidez/core/releases/tag/1.8.0) - 2024-01-23
+
+### Added
+
+- New payment events (#406)
+- Robots.txt from Magento config (#408)
+
+### Fixed
+
+- Use stores to get the token and mask on checkout success (#377)
+- Fix dusk tests (#405)
+
+## [1.7.1](https://github.com/rapidez/core/releases/tag/1.7.1) - 2024-01-05
+
+### Fixed
+
+- Exclude register themes when running from console (#401)
+
+## [1.7.0](https://github.com/rapidez/core/releases/tag/1.7.0) - 2024-01-04
+
+### Added
+
+- Payment icons (#374)
+- Store set event (#398)
+- Product review summary model (#399)
+- GraphQL components store prop (#400)
+
+### Fixed
+
+- Trigger mounted event for autocomplete to remove timing dependency (#395)
+
 ## [1.6.1](https://github.com/rapidez/core/releases/tag/1.6.1) - 2023-12-08
 
 ### Fixed